// src/app/routes/project/components/glass-layer/resize-layer/resize.interfaces.ts
/*
 * Copyright 2021 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { EDGE } from 'cd-utils/geometry';

export interface ILineAttr {
  edge: EDGE;
  x1: number;
  x2: number;
  y2: number;
  y1: number;
}

export interface IAnchorPoint {
  edge: EDGE;
  x: number;
  y: number;
}